+++ server/api/v1/clothing/style.go
@@ -2,21 +2,20 @@ package clothing
 
 import (
 	"github.com/flipped-aurora/gin-vue-admin/server/global"
-    "github.com/flipped-aurora/gin-vue-admin/server/model/clothing"
-    "github.com/flipped-aurora/gin-vue-admin/server/model/common/request"
-    clothingReq "github.com/flipped-aurora/gin-vue-admin/server/model/clothing/request"
-    "github.com/flipped-aurora/gin-vue-admin/server/model/common/response"
-    "github.com/flipped-aurora/gin-vue-admin/server/service"
-    "github.com/gin-gonic/gin"
-    "go.uber.org/zap"
-    "github.com/flipped-aurora/gin-vue-admin/server/utils"
+	"github.com/flipped-aurora/gin-vue-admin/server/model/clothing"
+	clothingReq "github.com/flipped-aurora/gin-vue-admin/server/model/clothing/request"
+	"github.com/flipped-aurora/gin-vue-admin/server/model/common/request"
+	"github.com/flipped-aurora/gin-vue-admin/server/model/common/response"
+	"github.com/flipped-aurora/gin-vue-admin/server/service"
+	"github.com/flipped-aurora/gin-vue-admin/server/utils"
+	"github.com/gin-gonic/gin"
+	"go.uber.org/zap"
 )
 
 type StyleApi struct {
 }
 
 var styleService = service.ServiceGroupApp.ClothingServiceGroup.StyleService
-
 
 // CreateStyle 创建Style
 // @Tags Style
@@ -34,9 +33,14 @@ func (styleApi *StyleApi) CreateStyle(c *gin.Context) {
 		response.FailWithMessage(err.Error(), c)
 		return
 	}
-    style.CreatedBy = utils.GetUserID(c)
+	style.CreatedBy = utils.GetUserID(c)
+	if !userRoleService.CheckManager(style.CreatedBy, style.CompanyID) {
+		global.GVA_LOG.Error("创建失败!", zap.Error(err))
+		response.FailWithMessage("权限不足", c)
+		return
+	}
 	if err := styleService.CreateStyle(&style); err != nil {
-        global.GVA_LOG.Error("创建失败!", zap.Error(err))
+		global.GVA_LOG.Error("创建失败!", zap.Error(err))
 		response.FailWithMessage("创建失败", c)
 	} else {
 		response.OkWithMessage("创建成功", c)
@@ -59,34 +63,30 @@ func (styleApi *StyleApi) DeleteStyle(c *gin.Context) {
 		response.FailWithMessage(err.Error(), c)
 		return
 	}
-    style.DeletedBy = utils.GetUserID(c)
+	style.CreatedBy = utils.GetUserID(c)
+	if !userRoleService.CheckManager(style.CreatedBy, style.CompanyID) {
+		global.GVA_LOG.Error("创建失败!", zap.Error(err))
+		response.FailWithMessage("权限不足", c)
+		return
+	}
 	if err := styleService.DeleteStyle(style); err != nil {
-        global.GVA_LOG.Error("删除失败!", zap.Error(err))
+		global.GVA_LOG.Error("删除失败!", zap.Error(err))
 		response.FailWithMessage("删除失败", c)
 	} else {
 		response.OkWithMessage("删除成功", c)
 	}
 }
 
-// DeleteStyleByIds 批量删除Style
-// @Tags Style
-// @Summary 批量删除Style
-// @Security ApiKeyAuth
-// @accept application/json
-// @Produce application/json
-// @Param data body request.IdsReq true "批量删除Style"
-// @Success 200 {string} string "{"success":true,"data":{},"msg":"批量删除成功"}"
-// @Router /style/deleteStyleByIds [delete]
 func (styleApi *StyleApi) DeleteStyleByIds(c *gin.Context) {
 	var IDS request.IdsReq
-    err := c.ShouldBindJSON(&IDS)
+	err := c.ShouldBindJSON(&IDS)
 	if err != nil {
 		response.FailWithMessage(err.Error(), c)
 		return
 	}
-    deletedBy := utils.GetUserID(c)
-	if err := styleService.DeleteStyleByIds(IDS,deletedBy); err != nil {
-        global.GVA_LOG.Error("批量删除失败!", zap.Error(err))
+	deletedBy := utils.GetUserID(c)
+	if err := styleService.DeleteStyleByIds(IDS, deletedBy); err != nil {
+		global.GVA_LOG.Error("批量删除失败!", zap.Error(err))
 		response.FailWithMessage("批量删除失败", c)
 	} else {
 		response.OkWithMessage("批量删除成功", c)
@@ -109,9 +109,14 @@ func (styleApi *StyleApi) UpdateStyle(c *gin.Context) {
 		response.FailWithMessage(err.Error(), c)
 		return
 	}
-    style.UpdatedBy = utils.GetUserID(c)
+	style.UpdatedBy = utils.GetUserID(c)
+	if !userRoleService.CheckManager(style.CreatedBy, style.CompanyID) {
+		global.GVA_LOG.Error("创建失败!", zap.Error(err))
+		response.FailWithMessage("权限不足", c)
+		return
+	}
 	if err := styleService.UpdateStyle(style); err != nil {
-        global.GVA_LOG.Error("更新失败!", zap.Error(err))
+		global.GVA_LOG.Error("更新失败!", zap.Error(err))
 		response.FailWithMessage("更新失败", c)
 	} else {
 		response.OkWithMessage("更新成功", c)
@@ -135,7 +140,7 @@ func (styleApi *StyleApi) FindStyle(c *gin.Context) {
 		return
 	}
 	if restyle, err := styleService.GetStyle(style.ID); err != nil {
-        global.GVA_LOG.Error("查询失败!", zap.Error(err))
+		global.GVA_LOG.Error("查询失败!", zap.Error(err))
 		response.FailWithMessage("查询失败", c)
 	} else {
 		response.OkWithData(gin.H{"restyle": restyle}, c)
@@ -159,14 +164,14 @@ func (styleApi *StyleApi) GetStyleList(c *gin.Context) {
 		return
 	}
 	if list, total, err := styleService.GetStyleInfoList(pageInfo); err != nil {
-	    global.GVA_LOG.Error("获取失败!", zap.Error(err))
-        response.FailWithMessage("获取失败", c)
-    } else {
-        response.OkWithDetailed(response.PageResult{
-            List:     list,
-            Total:    total,
-            Page:     pageInfo.Page,
-            PageSize: pageInfo.PageSize,
-        }, "获取成功", c)
-    }
+		global.GVA_LOG.Error("获取失败!", zap.Error(err))
+		response.FailWithMessage("获取失败", c)
+	} else {
+		response.OkWithDetailed(response.PageResult{
+			List:     list,
+			Total:    total,
+			Page:     pageInfo.Page,
+			PageSize: pageInfo.PageSize,
+		}, "获取成功", c)
+	}
 }
